Switch to concrete Success/Failure result classes (#18)
Data within an operation is now gathered on a separate `Data` object that’s passed to a concrete `Success`/`Failure` result instance on completion. Convenience data accessors can be defined on the `Data` object (via the renamed `.data_attr`) but remain available on the result using the magic of `#method_missing`.